add service load mechanism for TM (#938)
