+++ tm/src/main/java/io/seata/tm/api/DefaultGlobalTransaction.java
@@ -21,6 +21,7 @@ import io.seata.core.exception.TransactionException;
 import io.seata.core.model.GlobalStatus;
 import io.seata.core.model.TransactionManager;
 import io.seata.tm.DefaultTransactionManager;
+import io.seata.tm.TransactionManagerHolder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -60,7 +61,7 @@ public class DefaultGlobalTransaction implements GlobalTransaction {
      * @param role   the role
      */
     DefaultGlobalTransaction(String xid, GlobalStatus status, GlobalTransactionRole role) {
-        this.transactionManager = DefaultTransactionManager.get();
+        this.transactionManager = TransactionManagerHolder.get();
         this.xid = xid;
         this.status = status;
         this.role = role;
