+++ tm/src/main/resources/META-INF/services/io.seata.core.model.TransactionManager
@@ -1,0 +1,1 @@
+io.seata.tm.DefaultTransactionManager
